+++ plugins/v2m/v2/v2mconv.h
@@ -9,7 +9,7 @@ void ConvertV2M(const unsigned char *inptr, const int inlen, unsigned char **out
 
 extern const char * const v2mconv_errors[];
 
-unsigned long GetV2MPatchData(const unsigned char *inptr, const int inlen, 
+unsigned int GetV2MPatchData(const unsigned char *inptr, const int inlen,
 										 unsigned char **outptr, const unsigned char **patchmap);
 
 #endif
